requirements.txt
azure-cosmos==4.2.0
gunicorn==20.1.0
uvicorn==0.15.0
fastapi==0.70.0
pytest==6.2.5
pytest-cov==3.0.0
black==21.9b0